requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.13
langgraph==0.0.20

# LLM
ollama==0.1.6

# Database
sqlalchemy==2.0.25

# UI
rich==13.7.0
prompt-toolkit==3.0.43

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web UI (Phase 2 - optional for now)
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3

# Development
pytest==7.4.4
black==24.1.1
